@@ -74,8 +74,8 @@
   "private": true,
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.16.0",
-    "@ckeditor/ckeditor5-build-classic": "^19.0.0",
-    "@ckeditor/ckeditor5-react": "^2.1.0",
+    "@ckeditor/ckeditor5-build-classic": "^41.4.2",
+    "@ckeditor/ckeditor5-react": "^11.0.0",
     "@emotion/react": "^11.11.0",
     "@svgr/core": "^8.0.0",
     "@tanstack/react-table": "^8.20.5",
@@ -83,18 +83,19 @@
     "airbnb-prop-types": "^2.16.0",
     "highcharts": "^9.1.0",
     "highcharts-react-official": "^3.0.0",
+    "jotai": "^2.13.1",
     "lodash": "^4.17.20",
     "ogr2ogr": "^1.5.0",
     "ol": "^6.4.3",
-    "prop-types": "^15.7.2",
+    "prop-types": "^15.8.1",
     "quick-lint-js": "^3.2.0",
-    "react": "^16.8.0",
+    "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.1",
     "react-bootstrap": "^2.10.6",
     "react-data-table-component": "^7.6.2",
     "react-data-table-component-extensions": "^1.6.0",
     "react-datepicker": "^7.5.0",
-    "react-dom": "^16.8.0",
+    "react-dom": "^17.0.2",
     "react-grid-layout": "^1.3.0",
     "react-icons": "^5.5.0",
     "react-markdown": "^8.0.7",
@@ -128,7 +129,7 @@
     "tern": "^0.23.0",
     "vite": "^4.5.5",
     "vite-plugin-html": "^3.2.2",
-    "vite-plugin-node-polyfills": "^0.23.0",
+    "vite-plugin-node-polyfills": "^0.24.0",
     "vite-plugin-svgr": "^4.3.0",
     "vitest": "^2.0.3",
     "webpack": "^5.69.1",
