@@ -186,14 +186,14 @@
                            (str/join "', '" (map #(clean-header-text % design-type) invalid-headers))
                            "'")))))
 
-(defn load-external-data! [project-id distribution file-name file-base64 design-type primary-key]
+(defn load-external-data! [project-id distribution file-name file-base64 design-type primary-key]  
   (when (#{"csv" "shp" "geojson"} distribution)
     (let [folder-name (str tmp-dir "/ceo-tmp-" project-id "/")
           saved-file  (pu/write-file-part-base64 file-name
                                                  file-base64
                                                  folder-name
                                                  (str "project-" project-id "-" design-type))]
-      (pu/try-catch-throw #(let [[headers body] (get-file-data distribution design-type saved-file folder-name)]
+      (pu/try-catch-throw #(let [[headers body] (get-file-data distribution design-type saved-file folder-name)]                             
                              (when-not (seq body)
                                (pu/init-throw  (str "The " design-type " file contains no rows of data.")))
 
