@@ -6,7 +6,8 @@
             [ring.util.codec                      :refer [url-encode]]
             [ring.util.response                   :refer [redirect]]
             [triangulum.config                    :refer [get-config]]
-            [triangulum.response                  :refer [no-cross-traffic?]]
+            [triangulum.database                  :refer [call-sql]]
+            [triangulum.response                  :refer [no-cross-traffic? data-response]]
             [triangulum.type-conversion           :refer [val->int]]))
 
 
@@ -37,3 +38,16 @@
                      full-url
                      "&flash_message=You must login to see "
                      full-url)))))
+
+
+(defn crumb-data [{:keys [params session]}]
+  (let [user-id        (:userId session -1)
+        project        (:project params (:projectId params))
+        project        (when-let [project-id (val->int project)]
+                         (-> (call-sql "select_project_by_id" project-id)
+                             first :name))
+        institution    (when-let [inst-id (-> params :institution val->int)]
+                         (-> (call-sql "select_institution_by_id" inst-id user-id)
+                             first :name))]
+    (data-response {:project project
+                    :institution institution})))
