@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import { FormLayout, SectionBlock, StatsCell, StatsRow } from "./components/FormComponents";
-import { NavigationBar } from "./components/PageComponents";
+import { NavigationBar, BreadCrumbs } from "./components/PageComponents";
 import Modal from "./components/Modal";
 
 function Account(props) {
@@ -34,7 +34,7 @@ class UserStats extends React.Component {
   getUserStats = () => {
     fetch("/get-user-stats?accountId=" + this.props.accountId)
       .then((response) => (response.ok ? response.json() : Promise.reject(response)))
-      .then((stats) => this.setState({ stats }))
+      .then((stats) => this.setState((s)=>({... s,  stats })))
       .catch((response) => {
         console.log(response);
         this.setState ({modal: {alert: {alertType: "User Stats Alert",
@@ -76,7 +76,6 @@ class UserStats extends React.Component {
               </div>
             </div>
           </div>
-
           {perProject && (
             <div className="ProjectStats__user-table">
               <strong>User Stats:</strong>
@@ -86,7 +85,7 @@ class UserStats extends React.Component {
                   analysisTime={project.analysisAverage}
                   plots={project.plotCount}
                   title={`#${project.id} - ${project.name}`}
-                  titleHref={`/collection?projectId=${project.id}`}
+                  titleHref={`/collection?projectId=${project.id}&institutionId=${project.institution_rid}`}
                 />
               ))}
             </div>
@@ -216,6 +215,11 @@ class AccountForm extends React.Component {
 export function pageInit(params, session) {
   ReactDOM.render(
     <NavigationBar userId={session.userId} userName={session.userName} version={session.versionDeployed}>
+      <BreadCrumbs
+        crumbs={[
+          {display: "Account",
+           id:"account"}]}
+      />
       <Account
         accountId={parseInt(params.accountId || session.userId)}
         userId={session.userId}
