@@ -16,6 +16,7 @@ import fileIcon from "./fileIcon.svg";
 import helpIcon from "./helpIcon.svg";
 import infoIcon from "./infoIcon.svg";
 import leftArrowIcon from "./leftArrowIcon.svg";
+import leftArrowSlimIcon from "./leftArrowSlimIcon.svg";
 import leftDoubleIcon from "./leftDoubleIcon.svg";
 import lineStringIcon from "./lineStringIcon.svg";
 import linkIcon from "./linkIcon.svg";
@@ -37,6 +38,11 @@ import trashIcon from "./trashIcon.svg";
 import upArrowIcon from "./upArrowIcon.svg";
 import upCaretIcon from "./upCaretIcon.svg";
 import zoomInIcon from "./zoomInIcon.svg";
+import personIcon from "./personIcon.svg";
+import landscapeIcon from "./landscapeIcon.svg";
+import mapIcon from './mapIcon.svg';
+import folderIcon from './folderIcon.svg';
+import morevertIcon from './morevertIcon.svg';
 
 export const iconMap = {
   alert: alertIcon,
@@ -55,6 +61,7 @@ export const iconMap = {
   help: helpIcon,
   info: infoIcon,
   leftArrow: leftArrowIcon,
+  leftArrowSlim: leftArrowSlimIcon,
   leftDouble: leftDoubleIcon,
   lineString: lineStringIcon,
   link: linkIcon,
@@ -66,7 +73,7 @@ export const iconMap = {
   polygon: polygonIcon,
   question: questionIcon,
   refresh: refreshIcon,
-  rightArrow: rightArrowIcon,
+  rightArrow: rightArrowIcon,  
   rightCaret: rightCaretIcon,
   rightDouble: rightDoubleIcon,
   rule: ruleIcon,
@@ -76,6 +83,12 @@ export const iconMap = {
   upArrow: upArrowIcon,
   upCaret: upCaretIcon,
   zoomIn: zoomInIcon,
+  users: personIcon,
+  projects: landscapeIcon,
+  imagery: mapIcon,
+  folder: folderIcon,
+  moreVert: morevertIcon,
+
 };
 
 function SvgIcon({ icon, color, cursor, size, verticalAlign, transform }) {
