@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { NavigationBar } from "./components/PageComponents";
+import { NavigationBar, BreadCrumbs } from "./components/PageComponents";
 import Modal from "./components/Modal";
 
 class PasswordReset extends React.Component {
@@ -119,6 +119,10 @@ class PasswordReset extends React.Component {
 export function pageInit(params, session) {
   ReactDOM.render(
     <NavigationBar userId={-1} userName="" version={session.versionDeployed}>
+      <BreadCrumbs
+        crumbs={[{display: "Password Reset",
+                id:"pass-reset",}]}
+      />
       <PasswordReset email={session.email || ""} passwordResetKey={params.passwordResetKey || ""} />
     </NavigationBar>,
     document.getElementById("app")
