@@ -1,7 +1,7 @@
 import React from "react";
 import _ from "lodash";
 
-import AOIMap from "./AOIMap";
+import { AOIMap } from "./AOIMap";
 import SurveyQuestionsDesigner from "../survey/SurveyQuestionsDesigner";
 import SurveyCollectionPreview from "../survey/SurveyCollectionPreview";
 import SurveyRulesDesigner from "../survey/SurveyRulesDesigner";
@@ -195,7 +195,7 @@ export default class CreateProjectWizard extends React.Component {
       plotFileName: this.context.plotFileName,
       plotFileBase64: this.context.plotFileBase64,
       sampleFileName: this.context.sampleFileName,
-      sampleFileBase64: this.context.sampleFileBase64,
+      sampleFileBase64: this.context.sampleFileBase64
     });
 
     saveDraft = () => {
@@ -416,11 +416,11 @@ export default class CreateProjectWizard extends React.Component {
       useTemplatePlots,
       originalProject,
     } = this.context;
+    
     const totalPlots = this.getTotalPlots();
     const plotFileNeeded =
       !useTemplatePlots &&
       (projectId === -1 || plotDistribution !== originalProject.plotDistribution);
-    
     return getErrors({...this.context,
                       totalPlots: totalPlots,
                       plotFileNeeded: plotFileNeeded});
@@ -630,6 +630,7 @@ export default class CreateProjectWizard extends React.Component {
         {this.state.modal?.alert &&
          <Modal title={this.state.modal.alert.alertType}
                 onClose={()=>{this.setState({modal: null});
+                              this.state.modal.alert.onClose &&
                               this.state.modal.alert.onClose();}}>
            {this.state.modal.alert.alertMessage}
          </Modal>}
