@@ -1,4 +1,5 @@
-import React, { useContext, useState } from "react";
+import React, { useContext, useState, useEffect } from "react";
+
 import shp from "shpjs";
 import DatePicker from 'react-datepicker';
 
@@ -13,7 +14,6 @@ import { ProjectContext, plotLimit } from "./constants";
 import { mercator } from "../utils/mercator";
 import Modal from "../components/Modal";
 
-
 export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, projecType}){
   const context = useContext(ProjectContext);
   const {projectId, designSettings, newPlotShape, newPlotFileName, newPlotDistribution, setProjectDetails} = context;
@@ -24,12 +24,26 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
                                               latMin: "",
                                               lonMax: "",
                                               latMax: "",});
+  const [_totalPlots, setTotalPlots] = useState(totalPlots);
+
   const acceptedTypes = {
       csv: "text/csv",
       shp: "application/zip",
       geojson: "application/json",
     };
 
+  const setPlotDetails = (newDetail) => {
+    setPlotState({
+        lonMin: "",
+        latMin: "",
+        lonMax: "",
+        latMax: "",
+      });
+    setProjectDetails(
+      Object.assign(newDetail, { plots: [] }, { aoiFeatures: [], aoiFileName: "" },)
+    ); 
+  };
+v
   const checkPlotFile = (plotFileType, fileName, plotFileBase64) => {
     fetch("/check-plot-file", {
       method: "POST",
@@ -48,10 +62,11 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
       .then((data) => {
         const [[lonMin, latMin], [lonMax, latMax]] = data.fileBoundary;
         const aoiFeatures = [mercator.generateGeoJSON(latMin, latMax, lonMin, lonMax)];
+        setTotalPlots(_totalPlots + data.filePlotCount);
         setPlotState({lonMin, lonMax, latMin, latMax});
-        setProjectState({
+        setProjectState({          
           aoiFileName: fileName,
-          boundary: data.fileBoundary,
+          boundary: data.fileBoundary,          
           aoiFeatures});
         setNewPlotFileName(fileName);
         setProjectDetails({
@@ -63,7 +78,7 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
       );
   };
 
-  const renderFileInput = (fileType, append) => {
+  const renderFileInput = (fileType, append) => {    
     return(
       <div className="mb-3">
         <div style={{display: "flex"}}>
@@ -78,16 +93,16 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
               accept={acceptedTypes[fileType]}
               defaultValue=""
               id="new-plot-distribution-file"
-              onChange={(e) => {
+              onChange={(e) => {                
                 const file = e.target.files[0];
-                readFileAsBase64Url (file, (base64) => {
-                  checkPlotFile(fileType, file.name, base64);
-                  return setProjectDetails({
+                readFileAsBase64Url (file, (base64) => {                 
+                  checkPlotFile(fileType, file.name, base64);                  
+                  return setPlotDetails({
                     newPlotFileName: file.name,
                     newPlotFileBase64: base64,
                     append: true,
-                  });
-                });
+                  }); 
+                });		
               }}
               style={{ display: "none" }}
               type="file"
@@ -99,10 +114,10 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
           </label>
             </div>
       </div>
-    );
+    );  
   };
-
-  const renderPlotShape = () => {
+  
+  const renderPlotShape = () => {    
     return (
       <div className="form-group" style={{ display: "flex", flexDirection: "column" }}>
         <label>Plot shape</label>
@@ -112,7 +127,7 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
               checked={newPlotShape === "circle"}
               className="form-check-input"
               id="plot-shape-circle"
-              onChange={() => {setProjectDetails({ newPlotShape: "circle" });}}
+              onChange={() => {setPlotDetails({ newPlotShape: "circle" });}}
               type="radio"
             />
             <label className="form-check-label" htmlFor="plot-shape-circle">
@@ -124,7 +139,7 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
               checked={newPlotShape === "square"}
               className="form-check-input"
               id="plot-shape-square"
-              onChange={() => {setProjectDetails({ newPlotShape: "square"});}}
+              onChange={() => {setPlotDetails({ newPlotShape: "square"});}}
               type="radio"
             />
             <label className="form-check-label" htmlFor="plot-shape-square">
@@ -146,11 +161,11 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
         min="0"
         onChange={(e) =>{
           setNewPlotSize(e.target.value);
-          setProjectDetails({ [property]: Number(e.target.value) });}}
+          setPlotDetails({ [property]: Number(e.target.value) });}}
         step="1"
         type="number"
         value={newPlotSize}
-        disabled = {disabled}
+        disabled = {(context.availability === "published") || disabled}
       />
     </div>
   );
@@ -163,7 +178,7 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
         {renderFileInput("csv", append)}
         <div style={{ display: "flex" }}>
           <span className="mr-3">{renderPlotShape()}</span>
-          {renderLabeledInput(plotUnits, "newPlotSize")}
+          {renderLabeledInput(plotUnits, "plotSize")}
         </div>
       </div>
     );
@@ -200,13 +215,12 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
           <div className="form-group" style={{width:"fit-content"}}>
             <label>Spatial Distribution</label>
             <div style={{ position: "relative" }}>
-              {plotOptions[newPlotDistribution]?.alert &&
-               <p className="alert"> {plotOptions[newPlotDistribution]?.alert}</p>}
+              {plotOptions[newPlotDistribution].alert &&
+               <p className="alert"> {plotOptions[newPlotDistribution].alert}</p>}
               <select
                 className="form-control form-control-sm"
                 onChange={(e)=>{
-                  setProjectDetails({ newPlotDistribution: e.target.value});}
-		         }
+                  setPlotDetails({ newPlotDistribution: e.target.value});}}
                 value={newPlotDistribution}>
                 {Object.entries (plotOptions).map (([key, options]) => (
                   <option key={key} value={key}>
@@ -214,22 +228,22 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
                   </option>
                 ))}
               </select>
-            </div>
+            </div>            
           </div>
-          <p className="font-italic ml-2">{`- ${plotOptions[newPlotDistribution]?.description}`}</p>
+          <p className="font-italic ml-2">{`- ${plotOptions[newPlotDistribution].description}`}</p>          
         </div>
-        <div>
-          {plotOptions[newPlotDistribution]?.layout(newPlotDistribution, true)}</div>
+        <div>          
+          {plotOptions[newPlotDistribution].layout(newPlotDistribution, true)}</div>
 	<p
          className="font-italic ml-2 small"
          style={{
-           color: totalPlots > plotLimit ? "#8B0000" : "#006400",
+           color: _totalPlots > plotLimit ? "#8B0000" : "#006400",
            fontSize: "1rem",
            whiteSpace: "pre-line",
          }}
 	 >
-       {totalPlots > 0 &&
-              `This project will contain around ${formatNumberWithCommas(totalPlots)} plots.`}
+       {_totalPlots > 0 &&
+              `This project will contain around ${formatNumberWithCommas(_totalPlots)} plots.`}
             {totalPlots > 0 &&
               totalPlots > plotLimit &&
               `\n* The maximum allowed number for the selected plot distribution is ${formatNumberWithCommas(
@@ -241,94 +255,54 @@ export function NewPlotDesign ({aoiFeatures, institutionUserList, totalPlots, pr
   );
 };
 
-
-export class PlotDesign extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
+export function PlotDesign ({totalPlots, aoiFeatures}) {
+  const initState = {
       lonMin: "",
       latMin: "",
       lonMax: "",
       latMax: "",
       modal: null,
       plotIdList: [],
-    };
-  }
-  lastNumPlots = null;
-  lastCalculatedPlots = null;
-  lastCalcInputs = {};
+      totalPlots: "",
+  };
+  const [state, setState] = useState(initState);
 
-  componentDidMount() {
-    this.setCoordsFromBoundary();
-    if (this.context.numPlots) {
-      const plotIds = Array.from({ length: this.context.numPlots }, (_, i) => i + 1);
-      this.setState({ plotIdList: plotIds });
-    }
-  }
+  const context = useContext(ProjectContext);
 
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.aoiFeatures && prevProps.aoiFeatures !== this.props.aoiFeatures) {
-      this.setCoordsFromBoundary();
-    }
-
-    if (
-      this.context.type === "simplified" &&
-        (lonMin !== prevState.lonMin ||
-         latMin !== prevState.latMin ||
-         lonMax !== prevState.lonMax ||
-         latMax !== prevState.latMax)
-    ) {
-      this.setSimplifiedProjectDetails();
-    }
-    if (this.props.totalPlots &&
-        this.props.totalPlots !== prevProps.totalPlots &&
-        this.props.totalPlots <= 5000
-    ) {
-      const plotIds = Array.from({ length: this.props.totalPlots }, (_, i) => i + 1);
-      this.setState({ plotIdList: plotIds });
-    }
-  }
-
-  setCoordsFromBoundary = () => {
-    const { aoiFeatures } = this.props;
+  const lastNumPlots = null;
+  const lastCalculatedPlots = null;
+  const lastCalcInputs = {};
+  
+  const setCoordsFromBoundary = () => {
     if (aoiFeatures?.length === 1) {
       const boundaryExtent = mercator.parseGeoJson(aoiFeatures[0], false).getExtent();
-      this.setState({
-        lonMin: boundaryExtent[0],
-        latMin: boundaryExtent[1],
-        lonMax: boundaryExtent[2],
-        latMax: boundaryExtent[3],
-      });
+      setState((s)=> (
+        { ...s,
+          lonMin: boundaryExtent[0],
+          latMin: boundaryExtent[1],
+          lonMax: boundaryExtent[2],
+          latMax: boundaryExtent[3],
+        }));
     } else {
-      this.setState({
-        lonMin: "",
-        latMin: "",
-        lonMax: "",
-        latMax: "",
-      });
+      setState((s)=>(
+        { ...s,
+          lonMin: "",
+          latMin: "",
+          lonMax: "",
+          latMax: "",
+        }));
     }
   };
 
-  validBoundary = (latMin, latMax, lonMin, lonMax) =>
-    isNumber(latMin) &&
-    isNumber(latMax) &&
-    isNumber(lonMin) &&
-    isNumber(lonMax) &&
-    latMax > latMin &&
-    lonMax > lonMin;
+  const validBoundary = (latMin, latMax, lonMin, lonMax) =>
+        isNumber(latMin) &&
+        isNumber(latMax) &&
+        isNumber(lonMin) &&
+        isNumber(lonMax) &&
+        latMax > latMin &&
+        lonMax > lonMin;
 
-  updateBoundaryFromCoords = (newCoord) => {
-    this.setState(newCoord, () => {
-      const { latMin, latMax, lonMin, lonMax } = this.state;
-      if (this.validBoundary(latMin, latMax, lonMin, lonMax)) {
-        this.setPlotDetails({
-          aoiFeatures: [mercator.generateGeoJSON(latMin, latMax, lonMin, lonMax)],
-        });
-      }
-    });
-  };
-
-  setPlotDetails = (newDetail) => {
+  const setPlotDetails = (newDetail) => {
     const resetAOI = ["csv", "shp", "geojson"].includes(newDetail.plotDistribution);
     if (resetAOI) {
       this.setState({
@@ -338,18 +312,30 @@ export class PlotDesign extends React.Component {
         latMax: "",
       });
     }
-    this.context.setProjectDetails(
+    context.setProjectDetails(
       Object.assign(newDetail, { plots: [] }, resetAOI ? { aoiFeatures: [], aoiFileName: "" } : {})
     );
+    
   };
-
-  setSimplifiedProjectDetails = () => {
-    const { lonMin, lonMax, latMin } = this.state;
+  
+  const updateBoundaryFromCoords = (newCoord) => {
+    setState(newCoord, () => {
+      const { latMin, latMax, lonMin, lonMax } = state;
+      if (validBoundary(latMin, latMax, lonMin, lonMax)) {
+        setPlotDetails({
+          aoiFeatures: [mercator.generateGeoJSON(latMin, latMax, lonMin, lonMax)],
+        });
+      }
+    });
+  };
+  
+  const setSimplifiedProjectDetails = () => {
+    const { lonMin, lonMax, latMin } = state;
     const plotWidth = mercator.calculatePlotWidth(latMin, lonMin, lonMax);
-    const designSettings = this.context.designSettings;
-
-    this.context.setProjectDetails({
-      ...this.context,
+    const designSettings = context.designSettings;
+    
+    context.setProjectDetails({
+      ...context,
       sampleDistribution: "center",
       numPlots: 1,
       plotSize: Math.floor(plotWidth / 2),
@@ -358,40 +344,47 @@ export class PlotDesign extends React.Component {
       designSettings: {
         ...designSettings,
         sampleGeometries: {"lines": true, "points": true, "polygons": true}} });
-  }
+  };
 
-  /// Render Functions
 
-  renderLabeledInput = (label, property, disabled = false) => {
-    const { availability } = this.context;
-    return (
-      <div className="form-group" style={{ width: "fit-content" }}>
-        <label htmlFor={property}>{label}</label>
-        <input
-          className="form-control form-control-sm"
-          id={property}
-          min="0"
-          onChange={(e) => this.setPlotDetails({ [property]: Number(e.target.value) })}
-          step="1"
-          type="number"
-          value={this.context[property] || ""}
-          disabled = { (availability==="published") || disabled }
-        />
-      </div>
-    );
-  }
+  useEffect(()=> {    
+    setCoordsFromBoundary();
+    if (totalPlots &&
+        totalPlots <= 5000
+       ) {
+      const plotIds = Array.from({ length: totalPlots }, (_, i) => i + 1);
+      setState({ plotIdList: plotIds,
+                 totalPlots: plotIds.length});
+    }  
+  }, []);
 
-  renderShufflePlots = () => {
-    const { shufflePlots, availability } = this.context;
+  const renderLabeledInput = (label, property, disabled = false) => (
+    <div className="form-group" style={{ width: "fit-content" }}>
+      <label htmlFor={property}>{label}</label>
+      <input
+        className="form-control form-control-sm"
+        id={property}
+        min="0"
+        onChange={(e) => setPlotDetails({ [property]: Number(e.target.value) })}
+        step="1"
+        type="number"
+        value={context[property] || ""}
+        disabled = {disabled}
+      />
+    </div>
+  );
+
+  const renderShufflePlots = () => {
+    const { shufflePlots } = context;
     return (
       <div className="form-check">
         <input
           checked={shufflePlots}
           className="form-check-input"
           id="shufflePlots"
-          onChange={() => this.setPlotDetails({ shufflePlots: !shufflePlots })}
+          onChange={() => setPlotDetails({ shufflePlots: !shufflePlots })}
           type="checkbox"
-          disabled={availability==="published"}
+          disabled={context.availability==="published"}
         />
         <label className="form-check-label" htmlFor="shufflePlots">
           Shuffle plot order
@@ -400,8 +393,8 @@ export class PlotDesign extends React.Component {
     );
   };
 
-  renderPlotShape = () => {
-    const { plotShape, availability } = this.context;
+  const renderPlotShape = () => {
+    const { plotShape } = context;
     return (
       <div className="form-group" style={{ display: "flex", flexDirection: "column" }}>
         <label>Plot shape</label>
@@ -411,9 +404,9 @@ export class PlotDesign extends React.Component {
               checked={plotShape === "circle"}
               className="form-check-input"
               id="plot-shape-circle"
-              onChange={() => this.setPlotDetails({ plotShape: "circle" })}
+              onChange={() => setPlotDetails({ plotShape: "circle" })}
               type="radio"
-              disabled={availability==="published"}
+              disabled={context.availability==="published"}
             />
             <label className="form-check-label" htmlFor="plot-shape-circle">
               Circle
@@ -424,9 +417,9 @@ export class PlotDesign extends React.Component {
               checked={plotShape === "square"}
               className="form-check-input"
               id="plot-shape-square"
-              onChange={() => this.setPlotDetails({ plotShape: "square" })}
+              onChange={() => setPlotDetails({ plotShape: "square" })}
               type="radio"
-              disabled={availability==="published"}
+              disabled={context.availability==="published"}
             />
             <label className="form-check-label" htmlFor="plot-shape-square">
               Square
@@ -437,8 +430,8 @@ export class PlotDesign extends React.Component {
     );
   };
 
-  renderAOICoords = () => {
-    const { latMax, lonMin, lonMax, latMin, availability } = this.state;
+  const renderAOICoords = () => {
+    const { latMax, lonMin, lonMax, latMin , availability} = state;
     return (
       <div style={{ width: "20rem" }}>
         <div className="form-group ml-3">
@@ -449,7 +442,7 @@ export class PlotDesign extends React.Component {
                 max="90.0"
                 min="-90.0"
                 onChange={(e) =>
-                  this.updateBoundaryFromCoords({ latMax: parseFloat(e.target.value) })
+                  updateBoundaryFromCoords({ latMax: parseFloat(e.target.value) })
                 }
                 placeholder="North"
                 step="any"
@@ -466,7 +459,7 @@ export class PlotDesign extends React.Component {
                 max="180.0"
                 min="-180.0"
                 onChange={(e) =>
-                  this.updateBoundaryFromCoords({ lonMin: parseFloat(e.target.value) })
+                  updateBoundaryFromCoords({ lonMin: parseFloat(e.target.value) })
                 }
                 placeholder="West"
                 step="any"
@@ -481,7 +474,7 @@ export class PlotDesign extends React.Component {
                 max="180.0"
                 min="-180.0"
                 onChange={(e) =>
-                  this.updateBoundaryFromCoords({ lonMax: parseFloat(e.target.value) })
+                  updateBoundaryFromCoords({ lonMax: parseFloat(e.target.value) })
                 }
                 placeholder="East"
                 step="any"
@@ -498,7 +491,7 @@ export class PlotDesign extends React.Component {
                 max="90.0"
                 min="-90.0"
                 onChange={(e) =>
-                  this.updateBoundaryFromCoords({ latMin: parseFloat(e.target.value) })
+                  updateBoundaryFromCoords({ latMin: parseFloat(e.target.value) })
                 }
                 placeholder="South"
                 step="any"
@@ -513,28 +506,28 @@ export class PlotDesign extends React.Component {
     );
   };
 
-  loadGeoJson = (shpFile) => {
+  const loadGeoJson = (shpFile) => {
     try {
       shp(shpFile).then((g) => {
-        this.context.setProjectDetails({
+        context.setProjectDetails({
           aoiDataDebug: g.features,
           aoiFeatures: g.features.map((f) => f.geometry),
           aoiFileName: g.fileName,
         });
       });
     } catch {
-      this.setState ({modal: {alert: {alertType: "ShapeFile Error", alertMessage: "Unknown error loading shape file."}}});
+      setState ((s)=> ({...s, modal: {alert: {alertType: "ShapeFile Error", alertMessage: "Unknown error loading shape file."}}}));
     }
   };
 
-  renderBoundaryFileInput = () => {
-    const { aoiFileName, availability } = this.context;
+  const renderBoundaryFileInput = () => {
+    const { aoiFileName, availability } = context;
     return (
       <div className="d-flex flex-column">
-        {this.state.modal?.alert &&
-         <Modal title={this.state.modal.alert.alertType}
-                onClose={()=>{this.setState({modal: null});}}>
-           {this.state.modal.alert.alertMessage}
+        {state.modal?.alert &&
+         <Modal title={state.modal.alert.alertType}
+                onClose={()=>{setState((s)=> ({...s, modal: null}));}}>
+           {state.modal.alert.alertMessage}
          </Modal>}
 
         <div className="d-flex">
@@ -551,7 +544,7 @@ export class PlotDesign extends React.Component {
               id="project-boundary-file"
               onChange={(e) => {
                 const file = e.target.files[0];
-                readFileAsArrayBuffer(file, this.loadGeoJson);
+                readFileAsArrayBuffer(file, loadGeoJson);
               }}
               style={{ display: "none" }}
               type="file"
@@ -564,8 +557,8 @@ export class PlotDesign extends React.Component {
     );
   };
 
-  renderAOISelector = () => {
-    const { boundaryType, type, availability } = this.context;
+  const renderAOISelector = () => {
+    const { boundaryType, type, availability } = context;
     const boundaryOptions = [
       { value: "manual", label: "Input coordinates" },
       { value: "file", label: "Upload shp file" },
@@ -576,7 +569,7 @@ export class PlotDesign extends React.Component {
           <label>Boundary type</label>
           <select
             className="form-control form-control-sm"
-            onChange={(e) => this.setPlotDetails({ boundaryType: e.target.value })}
+            onChange={(e) => setPlotDetails({ boundaryType: e.target.value })}
             value={boundaryType}
             disabled={availability==="published"}
           >
@@ -587,12 +580,12 @@ export class PlotDesign extends React.Component {
             ))}
           </select>
         </div>
-        {boundaryType === "manual" ? this.renderAOICoords() : this.renderBoundaryFileInput()}
+        {boundaryType === "manual" ? renderAOICoords() : renderBoundaryFileInput()}
       </>
     );
   };
 
-  readFileAsArrayBuffer = (file) =>
+  const readFileAsArrayBuffer = (file) =>
     new Promise((res, rej) => {
       const r = new FileReader();
       r.onload = () => res(r.result);
@@ -600,7 +593,7 @@ export class PlotDesign extends React.Component {
       r.readAsArrayBuffer(file);
     });
 
-  readFileAsText = (file) =>
+  const readFileAsText = (file) =>
     new Promise((res, rej) => {
       const r = new FileReader();
       r.onload = () => res(r.result);
@@ -608,7 +601,7 @@ export class PlotDesign extends React.Component {
             r.readAsText(file);
     });
 
-  getPlotIdFromProps = (props = {}) => {
+  const getPlotIdFromProps = (props = {}) => {
     if (!props) return undefined;
     const keys = Object.keys(props);
     const k = keys.find((key) =>
@@ -618,25 +611,25 @@ export class PlotDesign extends React.Component {
   };
 
   // SHP file
-  parseZipShapefileToIds = async (file) => {
+  const parseZipShapefileToIds = async (file) => {
     const geojson = await shp(file);
     const features = geojson?.type === "FeatureCollection" ? geojson.features : [];
     return features
-      .map((f) => this.getPlotIdFromProps(f?.properties))
+      .map((f) => getPlotIdFromProps(f?.properties))
       .filter((v) => v != null);
   };
 
   // GeoJSON / JSON
-  parseGeoJSONToIds = (text) => {
+  const parseGeoJSONToIds = (text) => {
     const data = JSON.parse(text);
     const features = data?.type === "FeatureCollection" ? data.features : [];
     return features
-      .map((f) => this.getPlotIdFromProps(f?.properties))
+      .map((f) => getPlotIdFromProps(f?.properties))
       .filter((v) => v != null);
   };
 
   // CSV (first column = plotid)
-  parseCsvToIds = (text) => {
+  const parseCsvToIds = (text) => {
     const lines = text
           .split(/\r?\n/)
           .map((l) => l.trim())
@@ -658,29 +651,29 @@ export class PlotDesign extends React.Component {
     return ids;
   };
 
-  readPlotIdsFromFile = async (file) => {
+  const readPlotIdsFromFile = async (file) => {
     const name = (file?.name || "").toLowerCase();
 
     if (name.endsWith(".zip")) {
-      const ab = await this.readFileAsArrayBuffer(file);
-      return await this.parseZipShapefileToIds(ab);
+      const ab = await readFileAsArrayBuffer(file);
+      return await parseZipShapefileToIds(ab);
     }
 
     if (name.endsWith(".geojson") || name.endsWith(".json")) {
-      const text = await this.readFileAsText(file);
-      return this.parseGeoJSONToIds(text);
+      const text = await readFileAsText(file);
+      return parseGeoJSONToIds(text);
     }
 
     if (name.endsWith(".csv")) {
-      const text = await this.readFileAsText(file);
-      return this.parseCsvToIds(text);
+      const text = await readFileAsText(file);
+      return parseCsvToIds(text);
     }
 
     throw new Error("Unsupported file type. Please upload .zip (shp), .geojson/.json, or .csv.");
   };
 
-  checkPlotFile = (plotFileType, plotFileName, plotFileBase64) => {
-    const { projectId, designSettings } = this.context;
+  const checkPlotFile = (plotFileType, plotFileName, plotFileBase64) => {
+    const { projectId, designSettings } = context;
     fetch("/check-plot-file", {
       method: "POST",
       headers: {
@@ -695,17 +688,17 @@ export class PlotDesign extends React.Component {
       }),
     })
       .then((response) => (response.ok ? response.json() : Promise.reject(response)))
-      .then((data) =>
-        this.context.setProjectDetails({
+      .then((data) => {
+        setState((s) => ({... s, totalPlots: data.filePlotCount}));
+        context.setProjectDetails({
           designSettings: { ...designSettings,
                             userAssignment: data.userAssignment,
                             qaqcAssignment: data.qaqcAssignment}
-        })
-      );
-  }
+        });
+      });
+  };
 
-  renderFileInput = (fileType, append) => {
-    const { availability } = this.context;
+  const renderFileInput = (fileType, append) => {
     const acceptedTypes = {
       csv: "text/csv",
       shp: "application/zip",
@@ -728,12 +721,12 @@ export class PlotDesign extends React.Component {
               id="plot-distribution-file"
               onChange={(e) => {
                 const file = e.target.files[0];
-                this.readPlotIdsFromFile(file).then((plotIds) =>
-                  this.setState({ plotIdList: plotIds})
+                readPlotIdsFromFile(file).then((plotIds) =>
+                  setState((s) => ({...s,  plotIdList: plotIds}))
                 );
                 readFileAsBase64Url(file, (base64) => {
-                  this.checkPlotFile(fileType, file.name, base64);
-                  return this.setPlotDetails({
+                  checkPlotFile(fileType, file.name, base64);
+                  return setPlotDetails({
                     plotFileName: file.name,
                     plotFileBase64: base64,
                     append: false
@@ -742,12 +735,11 @@ export class PlotDesign extends React.Component {
               }}
               style={{ display: "none" }}
               type="file"
-              disabled={availability==="published"}
             />
           </label>
           <label className="ml-3 text-nowrap">
             File:{" "}
-            {this.context.plotFileName || (this.context.projectId > 0 ? "Use existing data" : "None")}
+            {context.plotFileName || (context.projectId > 0 ? "Use existing data" : "None")}
           </label>
         </div>
         <a
@@ -759,27 +751,27 @@ export class PlotDesign extends React.Component {
         </a>
       </div>
     );
-  }
+  };
   
-  renderCSV = (fileType, append) => {
-    const { plotShape } = this.context;
+  const renderCSV = (fileType, append) => {
+    const { plotShape } = context;
     const plotUnits = plotShape === "circle" ? "Plot diameter (m)" : "Plot width (m)";
     return (
       <div style={{ display: "flex", flexDirection: "column" }}>
-        {this.renderFileInput("csv", append)}
+        {renderFileInput("csv", append)}
         <div style={{ display: "flex" }}>
-          <span className="mr-3">{this.renderPlotShape()}</span>
-          {this.renderLabeledInput(plotUnits, "plotSize")}
+          <span className="mr-3">{renderPlotShape()}</span>
+          {renderLabeledInput(plotUnits, "plotSize")}
         </div>
       </div>
     );
   };
 
-  renderStrataRow = (feature, idx) => (
+  const renderStrataRow = (feature, idx) => (
     <div
       key={idx}
-      onMouseEnter={() => this.setPlotDetails({ selectedStrata: idx })}
-      onMouseLeave={() => this.setPlotDetails({ selectedStrata: -1 })}
+      onMouseEnter={() => setPlotDetails({ selectedStrata: idx })}
+      onMouseLeave={() => setPlotDetails({ selectedStrata: -1 })}
       style={{
         background: "rgba(255,255,255,0.5)",
         border: "1px solid black",
@@ -794,229 +786,229 @@ export class PlotDesign extends React.Component {
     </div>
   );
 
-  renderRandom = () => {
-    const { aoiFeatures, plotShape, type } = this.context;
+  const renderRandom = () => {
+    const { aoiFeatures, plotShape, type } = context;
     const disabled = type === "simplified";
     const plotUnits = plotShape === "circle" ? "Plot diameter (m)" : "Plot width (m)";
     return (
       <div>
-        {this.renderAOISelector()}
+        {renderAOISelector()}
         <label>Plot properties</label>
-        {aoiFeatures.map(this.renderStrataRow)}
+        {aoiFeatures.map(renderStrataRow)}
         <div className="d-flex">
-          {this.renderLabeledInput("Number of plots", "numPlots", disabled)}
-          <span className="mx-3">{this.renderPlotShape()}</span>
-          {this.renderLabeledInput(plotUnits, "plotSize")}
+          {renderLabeledInput("Number of plots", "numPlots", disabled)}
+          <span className="mx-3">{renderPlotShape()}</span>
+          {renderLabeledInput(plotUnits, "plotSize")}
         </div>
       </div>
     );
   };
 
-  renderGridded = () => {
-    const { aoiFeatures, plotShape } = this.context;
+  const renderGridded = () => {
+    const { aoiFeatures, plotShape } = context;
     const plotUnits = plotShape === "circle" ? "Plot diameter (m)" : "Plot width (m)";
     return (
       <div>
-        {this.renderAOISelector()}
-        {aoiFeatures.map(this.renderStrataRow)}
+        {renderAOISelector()}
+        {aoiFeatures.map(renderStrataRow)}
         <div className="d-flex">
-          {this.renderLabeledInput("Plot spacing (m)", "plotSpacing")}
-          <span className="mx-3">{this.renderPlotShape()}</span>
-          {this.renderLabeledInput(plotUnits, "plotSize")}
+          {renderLabeledInput("Plot spacing (m)", "plotSpacing")}
+          <span className="mx-3">{renderPlotShape()}</span>
+          {renderLabeledInput(plotUnits, "plotSize")}
         </div>
-        {this.renderShufflePlots()}
+        {renderShufflePlots()}
       </div>
     );
   };
 
-  renderSimplifiedSelector = () => {
-    const { aoiFeatures } = this.context;
+  const renderSimplifiedSelector = () => {
+    const { aoiFeatures } = context;
     return (
       <div>
-        {this.renderAOISelector()}
-        {aoiFeatures.map(this.renderStrataRow)}
+        {renderAOISelector()}
+        {aoiFeatures.map(renderStrataRow)}
       </div>
     );
   };
+  
+  const { plotDistribution, designSettings } = context;
+  
+  const plotOptions = {
+    random: {
+      display: "Random",
+      description: "Plot centers will be randomly distributed within the project boundary.",
+      layout: renderRandom,
+    },
+    gridded: {
+      display: "Gridded",
+      description:
+      "Plot centers will be arranged on a grid within the AOI using the plot spacing selected below.",
+      layout: renderGridded,
+    },
+    csv: {
+      display: "CSV File",
+      description:
+      "Specify your own plot centers by uploading a CSV with these fields: LON,LAT,PLOTID. Each plot center must have a unique PLOTID value.",
+      layout: renderCSV,
+    },
+    shp: {
+      display: "SHP File",
+      alert: "CEO may overestimate the number of project plots when using a ShapeFile.",
+      description:
+      "Specify your own plot boundaries by uploading a zipped Shapefile (containing SHP, SHX, DBF, and PRJ files) of polygon features. Each feature must have a unique PLOTID value.",
+      layout: renderFileInput,
+    },
+    geojson: {
+      display: "GeoJSON File",
+      alert: "CEO may overestimate the number of project plots when using a ShapeFile.",
+      description:
+      "Specify your own plot boundaries by uploading a GeoJSON file of polygon features. Each feature must have a unique PLOTID value in the properties map.",
+      layout: renderFileInput,
+    },
+  };
 
-  render() {
-    const { plotDistribution, designSettings } = this.context;
-    const { totalPlots } = this.props;
-    const plotOptions = {
-      random: {
-        display: "Random",
-        description: "Plot centers will be randomly distributed within the project boundary.",
-        layout: this.renderRandom,
-      },
-      gridded: {
-        display: "Gridded",
-        description:
-          "Plot centers will be arranged on a grid within the AOI using the plot spacing selected below.",
-        layout: this.renderGridded,
-      },
-      csv: {
-        display: "CSV File",
-        description:
-          "Specify your own plot centers by uploading a CSV with these fields: LON,LAT,PLOTID. Each plot center must have a unique PLOTID value.",
-        layout: this.renderCSV,
-      },
-      shp: {
-        display: "SHP File",
-        alert: "CEO may overestimate the number of project plots when using a ShapeFile.",
-        description:
-          "Specify your own plot boundaries by uploading a zipped Shapefile (containing SHP, SHX, DBF, and PRJ files) of polygon features. Each feature must have a unique PLOTID value.",
-        layout: this.renderFileInput,
-      },
-      geojson: {
-        display: "GeoJSON File",
-        alert: "CEO may overestimate the number of project plots when using a ShapeFile.",
-        description:
-          "Specify your own plot boundaries by uploading a GeoJSON file of polygon features. Each feature must have a unique PLOTID value in the properties map.",
-        layout: this.renderFileInput,
-      },
-    };
+  const simplifiedPlotOptions = {
+    ...filterObject(
+      plotOptions,
+      ([_id, item]) => item.display !== "Random" && item.display !== "Gridded"
+    ),
+    simplified: {
+      display: "Simplified AOI",
+      description: "Use the map preview or the coordinate inputs to create the project plot",
+      layout: renderSimplifiedSelector,
+    },
+  };
 
-    const simplifiedPlotOptions = {
-      ...filterObject(
-        plotOptions,
-        ([_id, item]) => item.display !== "Random" && item.display !== "Gridded"
-      ),
-      simplified: {
-        display: "Simplified AOI",
-        description: "Use the map preview or the coordinate inputs to create the project plot",
-        layout: this.renderSimplifiedSelector(),
-      },
-    };
+  const spatialDistributionOptions = context.type === 'simplified' ? simplifiedPlotOptions : plotOptions;
 
-    const spatialDistributionOptions = this.context.type === 'simplified' ? simplifiedPlotOptions : plotOptions;
-
-    return (
-      <div id="plot-design">
-        <h3 className="mb-3">Plot Generation</h3>
-        <div className="ml-3">
-          <div className="d-flex flex-column">
-            <div className="form-group" style={{ width: "fit-content" }}>
-              <label>Spatial distribution</label>
-              <div style={{position: "relative"}}>
-              {spatialDistributionOptions[plotDistribution]?.alert &&
-             <p className="alert">- {spatialDistributionOptions[plotDistribution]?.alert}</p>}
+  return (
+    <div id="plot-design">
+      <h3 className="mb-3">Plot Generation</h3>
+      <div className="ml-3">
+        <div className="d-flex flex-column">
+          <div className="form-group" style={{ width: "fit-content" }}>
+            <label>Spatial distribution</label>
+            <div style={{position: "relative"}}>
+              {spatialDistributionOptions[plotDistribution].alert &&
+               <p className="alert">- {spatialDistributionOptions[plotDistribution].alert}</p>}
               <select
                 className="form-control form-control-sm"
                 onChange={(e) =>
                   designSettings?.userAssignment?.userMethod === "file" ?
-                    this.setPlotDetails({
+                    setPlotDetails({
                       plotDistribution: e.target.value,
                       designSettings: { ...designSettings,
                                         userAssignment: {userMethod: "none", users: [], percents: []},
                                         qaqcAssignment: {qaqcMethod: "none", smes: [], overlap: 0}}})
-                    : this.setPlotDetails({ plotDistribution: e.target.value })
+                    : setPlotDetails({ plotDistribution: e.target.value })
                 }
                 value={plotDistribution}
-                disabled={this.context?.availability==="published"}
               >
                 {Object.entries(spatialDistributionOptions).map(([key, options]) => (
                   <option key={key} value={key}>
                     {options.display}
                   </option>
                 ))}
-      </select>
-      </div>
+              </select>
             </div>
-            <p className="font-italic ml-2">{`- ${spatialDistributionOptions[plotDistribution]?.description}`}</p>
           </div>
-          <div>{spatialDistributionOptions[plotDistribution]?.layout(plotDistribution, false)}</div>
-          <p
-            className="font-italic ml-2 small"
-            style={{
-              color: totalPlots > plotLimit ? "#8B0000" : "#006400",
-              fontSize: "1rem",
-              whiteSpace: "pre-line",
-            }}
-          >
-            {totalPlots > 0 &&
-             `This project will contain around ${formatNumberWithCommas(totalPlots)} plots.`}
-            {totalPlots > 0 &&
-             totalPlots > plotLimit &&
-             `\n* The maximum allowed number for the selected plot distribution is ${formatNumberWithCommas(
+          <p className="font-italic ml-2">{`- ${spatialDistributionOptions[plotDistribution].description}`}</p>
+          
+        </div>
+        <div>{spatialDistributionOptions[plotDistribution].layout(plotDistribution, false)}</div>
+        <p
+          className="font-italic ml-2 small"
+          style={{
+            color: state.totalPlots > plotLimit ? "#8B0000" : "#006400",
+            fontSize: "1rem",
+            whiteSpace: "pre-line",
+          }}
+        >
+          {state.totalPlots > 0 &&
+           `This project will contain around ${formatNumberWithCommas(totalPlots)} plots.`}
+          {state.totalPlots > 0 &&
+           state.totalPlots > plotLimit &&
+           `\n* The maximum allowed number for the selected plot distribution is ${formatNumberWithCommas(
                 plotLimit
               )}.`}
-          </p>
-          <h3 className="mb-3">Plot Similarity Configuration</h3>
-          <div className="form-check">
-            <input
-              checked={this.context.projectOptions.plotSimilarity}
-              className="form-check-input"
-              id="similarPlots"
-              onChange={() =>
-                this.context.setProjectDetails({
-                  projectOptions: { ...this.context.projectOptions, plotSimilarity: !this.context.projectOptions.plotSimilarity },
-                })
-              }
-              type="checkbox"
-            />
-            <label className="form-check-label" htmlFor="similarPlots">
-              Enable navigation by similarity
-            </label>
-            {this.context.projectOptions.plotSimilarity ? (
-              <>
-                <div className="form-group">
-                  <label htmlFor="referencePlotId"> Reference plot ID: {"  "}</label>
-                  <select
-                    id="referencePlotId"
-                    value={this.context.plotSimilarityDetails?.referencePlotId || ""}
-                    onChange={(e) =>
-                      this.context.setProjectDetails({
-                        plotSimilarityDetails: {
-                          ...this.context.plotSimilarityDetails,
-                          referencePlotId: e.target.value,
-                        },
-                      })
-                    }
-                  >
-                    <option value="" disabled>
-                      Select a plot ID
-                    </option>
-                    {this.state.plotIdList?.map((id) => (
-                      <option key={id} value={id}>
-                        {id}
+        </p>
+        {(context.type != "simplified") && (
+          <>
+            <h3 className="mb-3">Plot Similarity Configuration</h3>
+            <div className="form-check">
+              <input
+                checked={context.projectOptions.plotSimilarity}
+                className="form-check-input"
+                id="similarPlots"
+                onChange={() =>
+                  context.setProjectDetails({
+                    projectOptions: { ...context.projectOptions, plotSimilarity: !context.projectOptions.plotSimilarity },
+                  })
+                }
+                type="checkbox"
+              />
+              <label className="form-check-label" htmlFor="similarPlots">
+                Enable navigation by similarity
+              </label>
+              {context.projectOptions.plotSimilarity && (
+                <>
+                  <div className="form-group">
+                    <label htmlFor="referencePlotId"> Reference plot ID: {"  "}</label>
+                    <select
+                      id="referencePlotId"
+                      value={context.plotSimilarityDetails?.referencePlotId || ""}
+                      onChange={(e) =>
+                        context.setProjectDetails({
+                          plotSimilarityDetails: {
+                            ...context.plotSimilarityDetails,
+                            referencePlotId: e.target.value,
+                          },
+                        })
+                      }
+                    >
+                      <option value="" disabled>
+                        Select a plot ID
                       </option>
-                    ))}
-                  </select>
-                  <br/>
-                  <label htmlFor="year"> Year for comparison: {"  "} </label>
-                  <DatePicker
-                    selected={
-                      this.context.plotSimilarityDetails?.years?.[0]
-                        ? new Date(this.context.plotSimilarityDetails.years[0], 0, 1)
-                        : new Date(new Date().getFullYear() - 1, 0, 1)
-                    }
-                    onChange={(d) => {
-                      const year = d.getFullYear();
+                      {state.plotIdList?.map((id) => (
+                        <option key={id} value={id}>
+                          {id}
+                        </option>
+                      ))}
+                    </select>
+                    <br/>
+                    <label htmlFor="year"> Year for comparison: {"  "} </label>
+                    <DatePicker
+                      selected={
+                        context.plotSimilarityDetails?.years?.[0]
+                          ? new Date(context.plotSimilarityDetails.years[0], 0, 1)
+                          : new Date()
+                      }
+                      onChange={(d) => {
+                        const year = d.getFullYear();
 
-                      this.context.setProjectDetails({
-                        plotSimilarityDetails: {
-                          ...this.context.plotSimilarityDetails,
-                          years: [year]
-                        },
-                      });
-                    }}
-                    className="form-control"
-                    showYearPicker
-                    dateFormat="yyyy"
-                    minDate={new Date(2000, 0, 1)}
-                    maxDate={new Date(new Date().getFullYear() - 1, 0, 1)}
-                  />
-                </div>
-              </>
-            ) : null}
-          </div>
-        </div>
-        <hr/>
+                        context.setProjectDetails({
+                          plotSimilarityDetails: {
+                            ...context.plotSimilarityDetails,
+                            years: [year]
+                          },
+                        });
+                      }}
+                      className="form-control"
+                      showYearPicker
+                      dateFormat="yyyy"
+                    />
+                  </div>
+                </>
+              )}
+            </div>
+          </>
+        )}
       </div>
-    );
-  }
-}
-PlotDesign.contextType = ProjectContext;
+      <hr/>
+    </div>
+  );
+};
+
 
 export function PlotDesignReview() {
   const { institutionImagery } = useContext(ProjectContext);
