@@ -34,11 +34,15 @@ export const blankProject = {
     showPlotInformation: false,
     collectConfidence: false,
     autoLaunchGeoDash: true,
-  },
+    plotSimilarity: false
+  },  
+  newPlotDistribution: "csv",
+  newPlotFileName: "",
+  newPlotBase64: "",
   plotDistribution: "random",
   imageryId: -1,
   aoiFeatures: [],
-  aoiFileName: "",
+  aoiFileName: "",  
   boundaryType: "manual",
   numPlots: "",
   plotSpacing: null,
@@ -53,13 +57,13 @@ export const blankProject = {
   surveyRules: [],
   templateProjectId: -1,
   append: false,
-  newPlotFileName: "",
   newPlotFileBase64: "",
-  newPlotDistribution: "csv",
   newPlotSize: "",
   newPlotShape: "square",
   useTemplateWidgets: false,
   useTemplatePlots: false,
   projectImageryList: [],
   plots: [],
+  plotSimilarityDetails: {referencePlotId: 1,
+                          years: [2024]}
 };
