@@ -34,8 +34,8 @@ export const customStyles = {
   table: {
     style: {
       maxHeight: "600px",
-      overflowY: "auto",
       height: "600px",
+      overflowY: "auto",
     },
   },
 };
