@@ -235,10 +235,19 @@ def degradationTileUrl(requestDict):
         }
     return values
 
+# ########## DynamicWorld ##########
+
+def dynamicWorld(requestDict):    
+    geometry = getDefault(requestDict, 'geometry')
+    startDate="2021-01-01"
+    endDate="2022-01-01"
+    visParams = safeParseJSON(getDefault(requestDict, 'visParams', {}))
+    return imageCollectionToMapId("GOOGLE/DYNAMICWORLD/V1", visParams, 'Mode', startDate, endDate)
+
 
 # ########## Stats ##########
 
 
 def statistics(requestDict):
-    values = getStatistics(getDefault(requestDict, 'extent', None))
-    return values
+    values = getStatistics(getDefault(requestDict, 'extent', None))    
+    return {url: values}
